Load the Leaflet score demo only in browsers
